Stabilize ML-KEM; bump to v1.12.0
