+++ aws-lc-rs/src/unstable.rs
@@ -8,6 +8,8 @@
 //! # ⚠️ Warning
 //! Features contained within this module, or child modules are subject to changes, relocation,
 //! or removal across minor releases, and thus are not subject to semantic versioning policies.
-
+#[deprecated(note = "use  `aws_lc_rs::kdf` instead")]
 pub mod kdf;
+
+#[deprecated(note = "use `aws_lc_rs::kem` instead")]
 pub mod kem;
